(DOCSP-26540): Streamline Sync configuration section (#231)
Continuation of (DOCSP-25115): De-emphasize PBS across docs #217
Create "Sync Settings" page that details all the settings available in Sync. (This would have been on the Enable Sync page, but the tab headings show up in the On This Page bar for some reason)
Merge "Alter Your Sync Configuration" with "Enable Sync" as the steps are more or less the same. Move the caveats from "Alter" to the corresponding sections of the "Settings" and "Permissions" pages.
Move "Optimize Sync Atlas Usage" and "Compact an Atlas Volume" pages to the more advanced "Go to Production" section
Various rewrites and touch-ups, technical clarifications, etc. with an emphasis on flexible sync usage

* Move settings to own page

* WIP

* Merge alter and enable pages

* Update sync rules and permissions

* Move advanced stuff to production section

* Update landing and getting started

* Update redirects

* Fix links

* Update source/sync/configure/sync-settings.txt

Co-authored-by: Dachary <dc@dacharycarey.com>

* Address comments

Co-authored-by: Dachary <dc@dacharycarey.com>